pyproject.toml
[tool.poetry]
name = "pac"
version = "0.1.0"
description = ""
authors = ["raisultan <ki.xbozz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
clean-text = {version = "0.6.0", extras = ["gpl"]}
openai = "1.13.3"
python-dotenv = "1.0.1"
instructor = "0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
